+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### New features
 - [Online search] Add method to retrieve a POI by Mapbox ID
 
+### Breaking changes
+- [CORE] Undo renaming of all Discover-related terms to Category
+
 ## 2.0.0-beta.4
 
 ### New features