+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog for the Mapbox Search SDK for Android
 
-## 1.2.0-rc.2
+## 1.2.0
 
 ### New features
 - [CORE] We have added support for Mapbox IDs in search results. A new field called mapboxId is added to BaseSearchResult, SearchResult, CategoryResult, and OfflineSearchResult. Mapbox ID is stable, and developers can store it for lookup purposes.
@@ -10,7 +10,8 @@
 - [Address Autofill, Place Autocomplete] Search is now a two-step action. It returns `Suggestions` (without the geo coordinate and routable points) at the first step and `Result` (with the geo coordinate and routable points) after the suggestion is selected. Thus, the `coordinate` field is no longer available in `Suggestion`.
 - [Address Autofill, Place Autocomplete] Added a separate function `reverse` for coordinate-based searches.
 
-
+### Bug fixes
+- [Offline search] Don't return results with the empty name
 
 ## 1.0.0-rc.8